+++ src/clip/AudioClipState.ts
@@ -4,7 +4,7 @@
  * Describes the different states available for an audio clip.
  */
 export enum AudioClipState {
-  Stopped = "STOPPED",
-  Playing = "PLAYING",
-  Paused = "PAUSED"
+  STOPPED = "STOPPED",
+  PLAYING = "PLAYING",
+  PAUSED = "PAUSED"
 }