+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.0.3 / 2020-01-30
+==================
+* [FEATURE] Added ability to pause, resume, and stop.
+* [FEATURE] Added adjustable volume to each audio clip.
+
 0.0.2 / 2020-01-29
 ==================
 * [FEATURE] Added the following methods to the base API: `getAudio`, `removeAudio`, `removeAllAudio`.]